+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - `install_theme()` may signal the error "The theme already exists" by mistake (thanks, @YizhouZheng, #230).
 
+- A warning will be issued if two versions of Hugo are found (thanks, @mingsnu, #235).
+
 # CHANGES IN blogdown VERSION 0.3
 
 ## MAJOR CHANGES